requirements.txt
streamlit
openai
pdfplumber
python-dotenv
